requirements.txt
Flask==2.3.2
obsws-python==1.7.2
python-dotenv==1.0.1
websockets==12.0
gunicorn==21.2.0
Flask-Cors==4.0.0
